requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Flask-SQLAlchemy==3.1.1
PyMuPDF==1.23.8
python-docx==0.8.11
pandas==2.1.4
openpyxl==3.1.2
xlsxwriter==3.1.9